1.Newtonsoft.Json 全部配置 2.修改EF执行Sql参数化

+++ aspnet-core/src/AGooday.AgPay.Application/Services/MchPayPassageService.cs
@@ -130,7 +130,7 @@ namespace AGooday.AgPay.Application.Services
                     if (payPassage != null)
                     {
                         item.PassageId = payPassage.Id;
-                        item.State = payPassage.State;
+                        item.State = (sbyte)payPassage.State;
                         item.Rate = payPassage.Rate * 100;
                     }
                 }
